backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.100

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# AWS & Cloud (botocore version resolved automatically by boto3)

# Data & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
prophet==1.1.5
joblib==1.4.0
shap==0.45.0

# HTTP & API
httpx==0.27.0
aiohttp==3.9.5
requests==2.31.0

# Validation & Config
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Monitoring
structlog==24.1.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6

# Utilities
pytz==2024.1
arrow==1.3.0
